+++ __init__.py
@@ -54,14 +54,13 @@ except NameError:
         link_with_MizoresCustomExporter,
         menu_edit_mesh_context,
         menu_object_context,
-        op_apply_modifiers,
-        op_assign_lr_shapekey_tag,
-        op_separate_lr_shapekey,
-        op_separate_lr_shapekey_all,
-        op_separate_lr_shapekey_all_tag_detect,
-        op_separate_shapekeys,
-        op_sideofactive_point,
         translations,
+    )
+    from BlenderAddon_ShapeKeysUtil.scripts.ops import (
+        op_apply_modifiers, op_separate_lr_shapekey,
+        op_sideofactive_point,
+        op_separate_lr_shapekey_all_tag_detect, op_assign_lr_shapekey_tag, op_separate_lr_shapekey_all,
+        op_separate_shapekeys,
     )
 
 files = [
@@ -80,13 +79,13 @@ files = [
     link_with_MizoresCustomExporter,
     menu_edit_mesh_context,
     menu_object_context,
-    operator_apply_modifiers,
-    operator_assign_lr_shapekey_tag,
-    operator_separate_lr_shapekey,
-    operator_separate_lr_shapekey_all,
-    operator_separate_lr_shapekey_all_tag_detect,
-    operator_separate_shapekeys,
-    operator_sideofactive_point,
+    op_apply_modifiers,
+    op_assign_lr_shapekey_tag,
+    op_separate_lr_shapekey,
+    op_separate_lr_shapekey_all,
+    op_separate_lr_shapekey_all_tag_detect,
+    op_separate_shapekeys,
+    op_sideofactive_point,
     translations,
 ]
