+++ scripts/menu_object_context.py
@@ -17,9 +17,8 @@
 # ##### END GPL LICENSE BLOCK #####
 
 import bpy
-from . import op_separate_shapekeys, op_apply_modifiers
-from . import op_separate_lr_shapekey_all, op_separate_lr_shapekey_all_tag_detect
-from . import op_separate_lr_shapekey, op_assign_lr_shapekey_tag
+from .ops import op_apply_modifiers, op_separate_lr_shapekey, op_separate_lr_shapekey_all_tag_detect, \
+    op_assign_lr_shapekey_tag, op_separate_lr_shapekey_all, op_separate_shapekeys
 
 
 # オブジェクトモード　Special → ShapeKeys Util を登録する
@@ -34,13 +33,15 @@ class VIEW3D_MT_object_specials_shapekeys_util(bpy.types.Menu):
 
     def draw(self, context):
         layout = self.layout
-        layout.operator(operator_apply_modifiers.OBJECT_OT_specials_shapekeys_util_apply_modifiers.bl_idname)
-        layout.operator(operator_separate_shapekeys.OBJECT_OT_specials_shapekeys_util_shapekeys_to_objects.bl_idname)
+        layout.operator(op_apply_modifiers.OBJECT_OT_specials_shapekeys_util_apply_modifiers.bl_idname)
+        layout.operator(op_separate_shapekeys.OBJECT_OT_specials_shapekeys_util_shapekeys_to_objects.bl_idname)
         layout.separator()
-        layout.operator(operator_separate_lr_shapekey.OBJECT_OT_specials_shapekeys_util_separate_lr_shapekey.bl_idname)
-        layout.operator(operator_separate_lr_shapekey_all.OBJECT_OT_specials_shapekeys_util_separate_lr_shapekey_all.bl_idname)
-        layout.operator(operator_separate_lr_shapekey_all_tag_detect.OBJECT_OT_specials_shapekeys_util_separate_lr_shapekey_all_tag_detect.bl_idname)
-        layout.operator(operator_assign_lr_shapekey_tag.OBJECT_OT_specials_shapekeys_util_assign_lr_shapekey_tag.bl_idname)
+        layout.operator(op_separate_lr_shapekey.OBJECT_OT_specials_shapekeys_util_separate_lr_shapekey.bl_idname)
+        layout.operator(
+            op_separate_lr_shapekey_all.OBJECT_OT_specials_shapekeys_util_separate_lr_shapekey_all.bl_idname)
+        layout.operator(
+            op_separate_lr_shapekey_all_tag_detect.OBJECT_OT_specials_shapekeys_util_separate_lr_shapekey_all_tag_detect.bl_idname)
+        layout.operator(op_assign_lr_shapekey_tag.OBJECT_OT_specials_shapekeys_util_assign_lr_shapekey_tag.bl_idname)
 
 
 def register():
