+++ scripts/ops/op_apply_modifiers.py
@@ -18,7 +18,7 @@
 
 import bpy
 from bpy.props import BoolProperty
-from . import consts, func_utils, func_apply_modifiers_with_shapekeys
+from .. import consts, func_utils, func_apply_modifiers_with_shapekeys
 
 
 class OBJECT_OT_specials_shapekeys_util_apply_modifiers(bpy.types.Operator):
