+++ scripts/translations.py
@@ -19,15 +19,15 @@
 import bpy
 from . import consts, func_package_utils
 from .consts import DESC
-from .op_separate_lr_shapekey import OBJECT_OT_specials_shapekeys_util_separate_lr_shapekey as separate_lr_shapekey
-from .op_assign_lr_shapekey_tag import OBJECT_OT_specials_shapekeys_util_assign_lr_shapekey_tag as assign_lr_shapekey_tag
-from .op_separate_lr_shapekey_all import OBJECT_OT_specials_shapekeys_util_separate_lr_shapekey_all as separate_lr_shapekey_all
-from .op_separate_lr_shapekey_all_tag_detect import (
+from .ops.op_separate_lr_shapekey import OBJECT_OT_specials_shapekeys_util_separate_lr_shapekey as separate_lr_shapekey
+from .ops.op_assign_lr_shapekey_tag import OBJECT_OT_specials_shapekeys_util_assign_lr_shapekey_tag as assign_lr_shapekey_tag
+from .ops.op_separate_lr_shapekey_all import OBJECT_OT_specials_shapekeys_util_separate_lr_shapekey_all as separate_lr_shapekey_all
+from .ops.op_separate_lr_shapekey_all_tag_detect import (
     OBJECT_OT_specials_shapekeys_util_separate_lr_shapekey_all_tag_detect as separate_lr_shapekey_all_tag_detect
 )
-from .op_apply_modifiers import OBJECT_OT_specials_shapekeys_util_apply_modifiers as apply_modifiers
-from .op_separate_shapekeys import OBJECT_OT_specials_shapekeys_util_shapekeys_to_objects as shapekeys_to_objects
-from .op_sideofactive_point import MESH_OT_specials_shapekeys_util_sideofactive_point as sideofactive_point
+from .ops.op_apply_modifiers import OBJECT_OT_specials_shapekeys_util_apply_modifiers as apply_modifiers
+from .ops.op_separate_shapekeys import OBJECT_OT_specials_shapekeys_util_shapekeys_to_objects as shapekeys_to_objects
+from .ops.op_sideofactive_point import MESH_OT_specials_shapekeys_util_sideofactive_point as sideofactive_point
 
 
 translations_dict = {
